Reduce memory allocation by avoiding some useless calls to `ReplaceAllStringFunc()`.

This results in a nice speed up for most of the exported functions. 
This also remove some german tests that were failing because of missing entries in `"de"` dict.

**benchmark results:** 

```
benchmark                          old ns/op     new ns/op     delta
BenchmarkCellPhone-2               4151          4161          +0.24%
BenchmarkCity-2                    3668          3752          +2.29%
BenchmarkCityPrefix-2              381           182           -52.23%
BenchmarkCitySuffix-2              402           427           +6.22%
BenchmarkCmapnyBS-2                1659          579           -65.10%
BenchmarkCompagnyCatchPhrase-2     1651          632           -61.72%
BenchmarkCompagnyName-2            3830          3815          -0.39%
BenchmarkCompagnySuffix-2          389           407           +4.63%
BenchmarkCountry-2                 457           459           +0.44%
BenchmarkDomainName-2              6377          5662          -11.21%
BenchmarkDomainSuffix-2            407           140           -65.60%
BenchmarkDomainWord-2              5325          5473          +2.78%
BenchmarkEmail-2                   8330          7140          -14.29%
BenchmarkFirstName-2               397           118           -70.28%
BenchmarkFreeEmail-2               2410          1628          -32.45%
BenchmarkJobTitle-2                1405          573           -59.22%
BenchmarkLastName-2                411           126           -69.34%
BenchmarkName-2                    3719          3728          +0.24%
BenchmarkNamePrefix-2              383           120           -68.67%
BenchmarkNameSuffix-2              386           124           -67.88%
BenchmarkPhoneNumber-2             4490          4148          -7.62%
BenchmarkPostCode-2                2594          2393          -7.75%
BenchmarkSafeEmail-2               2080          1163          -44.09%
BenchmarkSecondaryAdress-2         1842          1747          -5.16%
BenchmarkState-2                   414           136           -67.15%
BenchmarkSateAbbr-2                366           157           -57.10%
BenchmarkStreetAdress-2            8292          8388          +1.16%
BenchmarkSteetName-2               3477          3521          +1.27%
BenchmarkStreeSuffix-2             409           143           -65.04%
BenchmarkUrl-2                     8151          7032          -13.73%
BenchmarkUserName-2                1924          1015          -47.25%

benchmark                          old allocs     new allocs     delta
BenchmarkCellPhone-2               18             17             -5.56%
BenchmarkCity-2                    10             10             +0.00%
BenchmarkCityPrefix-2              3              0              -100.00%
BenchmarkCitySuffix-2              3              3              +0.00%
BenchmarkCmapnyBS-2                11             1              -90.91%
BenchmarkCompagnyCatchPhrase-2     11             1              -90.91%
BenchmarkCompagnyName-2            11             11             +0.00%
BenchmarkCompagnySuffix-2          3              3              +0.00%
BenchmarkCountry-2                 3              3              +0.00%
BenchmarkDomainName-2              21             18             -14.29%
BenchmarkDomainSuffix-2            3              0              -100.00%
BenchmarkDomainWord-2              17             17             +0.00%
BenchmarkEmail-2                   35             23             -34.29%
BenchmarkFirstName-2               3              0              -100.00%
BenchmarkFreeEmail-2               17             5              -70.59%
BenchmarkJobTitle-2                10             1              -90.00%
BenchmarkLastName-2                3              0              -100.00%
BenchmarkName-2                    9              9              +0.00%
BenchmarkNamePrefix-2              3              0              -100.00%
BenchmarkNameSuffix-2              3              0              -100.00%
BenchmarkPhoneNumber-2             18             17             -5.56%
BenchmarkPostCode-2                12             12             +0.00%
BenchmarkSafeEmail-2               14             5              -64.29%
BenchmarkSecondaryAdress-2         11             11             +0.00%
BenchmarkState-2                   3              0              -100.00%
BenchmarkSateAbbr-2                3              0              -100.00%
BenchmarkStreetAdress-2            25             25             +0.00%
BenchmarkSteetName-2               9              9              +0.00%
BenchmarkStreeSuffix-2             3              0              -100.00%
BenchmarkUrl-2                     35             23             -34.29%
BenchmarkUserName-2                13             4              -69.23%

benchmark                          old bytes     new bytes     delta
BenchmarkCellPhone-2               286           270           -5.59%
BenchmarkCity-2                    195           195           +0.00%
BenchmarkCityPrefix-2              32            0             -100.00%
BenchmarkCitySuffix-2              32            32            +0.00%
BenchmarkCmapnyBS-2                209           37            -82.30%
BenchmarkCompagnyCatchPhrase-2     223           42            -81.17%
BenchmarkCompagnyName-2            237           237           +0.00%
BenchmarkCompagnySuffix-2          32            32            +0.00%
BenchmarkCountry-2                 44            44            +0.00%
BenchmarkDomainName-2              394           362           -8.12%
BenchmarkDomainSuffix-2            32            0             -100.00%
BenchmarkDomainWord-2              343           343           +0.00%
BenchmarkEmail-2                   618           488           -21.04%
BenchmarkFirstName-2               32            0             -100.00%
BenchmarkFreeEmail-2               269           123           -54.28%
BenchmarkJobTitle-2                148           34            -77.03%
BenchmarkLastName-2                34            0             -100.00%
BenchmarkName-2                    162           162           +0.00%
BenchmarkNamePrefix-2              32            0             -100.00%
BenchmarkNameSuffix-2              32            0             -100.00%
BenchmarkPhoneNumber-2             286           270           -5.59%
BenchmarkPostCode-2                143           144           +0.70%
BenchmarkSafeEmail-2               224           126           -43.75%
BenchmarkSecondaryAdress-2         135           135           +0.00%
BenchmarkState-2                   38            0             -100.00%
BenchmarkSateAbbr-2                32            0             -100.00%
BenchmarkStreetAdress-2            592           592           +0.00%
BenchmarkSteetName-2               187           187           +0.00%
BenchmarkStreeSuffix-2             32            0             -100.00%
BenchmarkUrl-2                     625           494           -20.96%
BenchmarkUserName-2                193           95            -50.78%
```